step 3: public manifestregistry types
