+++ crates/store/re_protos/src/lib.rs
@@ -35,12 +35,6 @@ mod v1alpha1 {
     #[path = "./rerun.sdk_comms.v1alpha1.rs"]
     pub mod rerun_sdk_comms_v1alpha1;
 
-    #[path = "./rerun.manifest_registry.v1alpha1.rs"]
-    pub mod rerun_manifest_registry_v1alpha1;
-
-    #[path = "./rerun.manifest_registry.v1alpha1.ext.rs"]
-    pub mod rerun_manifest_registry_v1alpha1_ext;
-
     #[path = "./rerun.frontend.v1alpha1.rs"]
     pub mod rerun_frontend_v1alpha1;
 
@@ -66,12 +60,12 @@ pub mod log_msg {
     }
 }
 
-pub mod manifest_registry {
+pub mod frontend {
     #[rustfmt::skip] // keep these constants single line for easy sorting
     pub mod v1alpha1 {
-        pub use crate::v1alpha1::rerun_manifest_registry_v1alpha1::*;
+        pub use crate::v1alpha1::rerun_frontend_v1alpha1::*;
         pub mod ext {
-            pub use crate::v1alpha1::rerun_manifest_registry_v1alpha1_ext::*;
+            pub use crate::v1alpha1::rerun_frontend_v1alpha1_ext::*;
         }
 
         /// `DatasetManifest` mandatory field names. All mandatory metadata fields are prefixed
@@ -82,15 +76,6 @@ pub mod manifest_registry {
         pub const DATASET_MANIFEST_REGISTRATION_TIME_FIELD_NAME: &str = "rerun_registration_time";
         pub const DATASET_MANIFEST_START_TIME_FIELD_NAME: &str = "rerun_start_time";
         pub const DATASET_MANIFEST_STORAGE_URL_FIELD_NAME: &str = "rerun_storage_url";
-    }
-}
-
-pub mod frontend {
-    pub mod v1alpha1 {
-        pub use crate::v1alpha1::rerun_frontend_v1alpha1::*;
-        pub mod ext {
-            pub use crate::v1alpha1::rerun_frontend_v1alpha1_ext::*;
-        }
     }
 }
