+++ crates/viewer/re_redap_browser/src/entries.rs
@@ -20,9 +20,7 @@ use re_protos::TypeConversionError;
 use re_protos::external::prost;
 use re_protos::external::prost::Name as _;
 use re_protos::frontend::v1alpha1::ext::{EntryDetails, TableEntry};
-use re_protos::frontend::v1alpha1::{
-    EntryFilter, EntryKind, FindEntriesRequest, ext::DatasetEntry,
-};
+use re_protos::frontend::v1alpha1::{EntryFilter, EntryKind, ext::DatasetEntry};
 use re_sorbet::SorbetError;
 use re_ui::{Icon, icons};
 use re_viewer_context::AsyncRuntimeHandle;
