+++ crates/viewer/re_redap_browser/src/servers.rs
@@ -10,8 +10,8 @@ use re_auth::Jwt;
 use re_dataframe_ui::{ColumnBlueprint, default_display_name_for_column};
 use re_grpc_client::ConnectionRegistryHandle;
 use re_log_types::{EntityPathPart, EntryId};
+use re_protos::frontend::v1alpha1::DATASET_MANIFEST_ID_FIELD_NAME;
 use re_protos::frontend::v1alpha1::EntryKind;
-use re_protos::manifest_registry::v1alpha1::DATASET_MANIFEST_ID_FIELD_NAME;
 use re_sorbet::{BatchType, ColumnDescriptorRef};
 use re_ui::alert::Alert;
 use re_ui::{UiExt as _, icons};
