+++ rerun_py/src/catalog/mod.rs
@@ -68,7 +68,7 @@ enum PyVectorDistanceMetric {
     Hamming,
 }
 
-impl From<PyVectorDistanceMetric> for re_protos::manifest_registry::v1alpha1::VectorDistanceMetric {
+impl From<PyVectorDistanceMetric> for re_protos::frontend::v1alpha1::VectorDistanceMetric {
     fn from(metric: PyVectorDistanceMetric) -> Self {
         match metric {
             PyVectorDistanceMetric::L2 => Self::L2,
@@ -89,9 +89,7 @@ enum VectorDistanceMetricLike {
     CatchAll(String),
 }
 
-impl TryFrom<VectorDistanceMetricLike>
-    for re_protos::manifest_registry::v1alpha1::VectorDistanceMetric
-{
+impl TryFrom<VectorDistanceMetricLike> for re_protos::frontend::v1alpha1::VectorDistanceMetric {
     type Error = PyErr;
 
     fn try_from(metric: VectorDistanceMetricLike) -> Result<Self, PyErr> {
@@ -112,8 +110,7 @@ impl TryFrom<VectorDistanceMetricLike>
 
 impl From<PyVectorDistanceMetric> for i32 {
     fn from(metric: PyVectorDistanceMetric) -> Self {
-        let proto_typed =
-            re_protos::manifest_registry::v1alpha1::VectorDistanceMetric::from(metric);
+        let proto_typed = re_protos::frontend::v1alpha1::VectorDistanceMetric::from(metric);
 
         proto_typed as Self
     }
